add test when A layer depends on B and B depends on A
