+++ lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture2/project/domain/sample/DomainSecondClass.kt
@@ -1,5 +1,5 @@
 package com.lemonappdev.konsist.architecture.assertarchitecture.architecture2.project.domain.sample
 
-import com.lemonappdev.konsist.architecture.assertarchitecture.architecture2.project.domain.DomainFirstClass
+import com.lemonappdev.konsist.architecture.assertarchitecture.architecture3.project.domain.DomainFirstClass
 
 class DomainSecondClass(val sampleParameter: DomainFirstClass)
