+++ lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture2/project/presentation/sample/PresentationSecondClass.kt
@@ -1,5 +1,5 @@
 package com.lemonappdev.konsist.architecture.assertarchitecture.architecture2.project.presentation.sample
 
-import com.lemonappdev.konsist.architecture.assertarchitecture.architecture2.project.presentation.PresentationFirstClass
+import com.lemonappdev.konsist.architecture.assertarchitecture.architecture3.project.presentation.PresentationFirstClass
 
 class PresentationSecondClass(val sampleParameter: PresentationFirstClass)
