+++ lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture3/README.md
@@ -1,9 +1,10 @@
 # Architecture 3
 
-Test scenario where `Presentation` and `Data` layers depend on `Domain` layer:
+Test scenario where `Presentation` layer depends on `Domain` one and `Domain` layer depends on `Presentation` one:
 
 ```mermaid
 %%{init: {'theme': 'forest'}}%%
 flowchart LR
-    Presentation & Data --> Domain
+    Presentation --> Domain
+    Domain --> Presentation
 ```
