--- lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture3/project/data/DataFirstClass.kt
@@ -1,3 +1,0 @@
-package com.lemonappdev.konsist.architecture.assertarchitecture.architecture3.project.data
-
-class DataFirstClass
