--- lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture3/project/data/sample/DataSecondClass.kt
@@ -1,5 +1,0 @@
-package com.lemonappdev.konsist.architecture.assertarchitecture.architecture3.project.data.sample
-
-import com.lemonappdev.konsist.architecture.assertarchitecture.architecture3.project.data.DataFirstClass
-
-class DataSecondClass(val sampleParameter: DataFirstClass)
