+++ lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture3/project/domain/sample/DomainThirdClass.kt
@@ -1,0 +1,5 @@
+package com.lemonappdev.konsist.architecture.assertarchitecture.architecture3.project.domain.sample
+
+import com.lemonappdev.konsist.architecture.assertarchitecture.architecture3.project.presentation.PresentationFirstClass
+
+class DomainThirdClass(val sampleParameter: PresentationFirstClass)
