+++ lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture4/README.md
@@ -1,13 +1,9 @@
 # Architecture 4
 
-Test scenario where `Presentation` depends on `Application`, `Application` depends on `Domain` and `Infrastructure`,
-`Domain` depends on `Infrastructure`:
-
+Test scenario where `Presentation` and `Data` layers depend on `Domain` layer:
 
 ```mermaid
 %%{init: {'theme': 'forest'}}%%
 flowchart LR
-    Presentation --> Application
-    Application & Domain --> Infrastructure
-    Application --> Domain
+    Presentation & Data --> Domain
 ```
