--- lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture4/project/application/ApplicationFirstClass.kt
@@ -1,3 +1,0 @@
-package com.lemonappdev.konsist.architecture.assertarchitecture.architecture4.project.application
-
-class ApplicationFirstClass
