--- lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture4/project/application/sample/ApplicationFourthClass.kt
@@ -1,5 +1,0 @@
-package com.lemonappdev.konsist.architecture.assertarchitecture.architecture4.project.application.sample
-
-import com.lemonappdev.konsist.architecture.assertarchitecture.architecture4.project.infrastructure.InfrastructureFirstClass
-
-class ApplicationFourthClass(val sampleParameter: InfrastructureFirstClass)
