+++ lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture4/project/data/sample/DataSecondClass.kt
@@ -1,0 +1,5 @@
+package com.lemonappdev.konsist.architecture.assertarchitecture.architecture4.project.data.sample
+
+import com.lemonappdev.konsist.architecture.assertarchitecture.architecture4.project.data.DataFirstClass
+
+class DataSecondClass(val sampleParameter: DataFirstClass)
