+++ lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture4/project/data/sample/DataThirdClass.kt
@@ -1,0 +1,5 @@
+package com.lemonappdev.konsist.architecture.assertarchitecture.architecture4.project.data.sample
+
+import com.lemonappdev.konsist.architecture.assertarchitecture.architecture4.project.domain.DomainFirstClass
+
+class DataThirdClass(val sampleParameter: DomainFirstClass)
