--- lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture4/project/domain/sample/DomainThirdClass.kt
@@ -1,5 +1,0 @@
-package com.lemonappdev.konsist.architecture.assertarchitecture.architecture4.project.domain.sample
-
-import com.lemonappdev.konsist.architecture.assertarchitecture.architecture4.project.infrastructure.InfrastructureFirstClass
-
-class DomainThirdClass(val sampleParameter: InfrastructureFirstClass)
