--- lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture4/project/infrastructure/InfrastructureFirstClass.kt
@@ -1,3 +1,0 @@
-package com.lemonappdev.konsist.architecture.assertarchitecture.architecture4.project.infrastructure
-
-class InfrastructureFirstClass
