--- lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture4/project/infrastructure/sample/InfrastructureSecondClass.kt
@@ -1,5 +1,0 @@
-package com.lemonappdev.konsist.architecture.assertarchitecture.architecture4.project.infrastructure.sample
-
-import com.lemonappdev.konsist.architecture.assertarchitecture.architecture4.project.infrastructure.InfrastructureFirstClass
-
-class InfrastructureSecondClass(val sampleParameter: InfrastructureFirstClass)
