+++ lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture4/project/presentation/sample/PresentationThirdClass.kt
@@ -1,5 +1,5 @@
 package com.lemonappdev.konsist.architecture.assertarchitecture.architecture4.project.presentation.sample
 
-import com.lemonappdev.konsist.architecture.assertarchitecture.architecture4.project.application.ApplicationFirstClass
+import com.lemonappdev.konsist.architecture.assertarchitecture.architecture4.project.domain.DomainFirstClass
 
-class PresentationThirdClass(val sampleParameter: ApplicationFirstClass)
+class PresentationThirdClass(val sampleParameter: DomainFirstClass)
