+++ lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture5/README.md
@@ -1,3 +1,13 @@
 # Architecture 5
 
-Test scenario where `EmptyLayer` has no files inside.
+Test scenario where `Presentation` depends on `Application`, `Application` depends on `Domain` and `Infrastructure`,
+`Domain` depends on `Infrastructure`:
+
+
+```mermaid
+%%{init: {'theme': 'forest'}}%%
+flowchart LR
+    Presentation --> Application
+    Application & Domain --> Infrastructure
+    Application --> Domain
+```
