+++ lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture5/project/application/ApplicationFirstClass.kt
@@ -1,0 +1,3 @@
+package com.lemonappdev.konsist.architecture.assertarchitecture.architecture5.project.application
+
+class ApplicationFirstClass
