+++ lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture5/project/application/sample/ApplicationSecondClass.kt
@@ -1,0 +1,5 @@
+package com.lemonappdev.konsist.architecture.assertarchitecture.architecture5.project.application.sample
+
+import com.lemonappdev.konsist.architecture.assertarchitecture.architecture5.project.application.ApplicationFirstClass
+
+class ApplicationSecondClass(val sampleParameter: ApplicationFirstClass)
