+++ lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture5/project/application/sample/ApplicationThirdClass.kt
@@ -1,0 +1,5 @@
+package com.lemonappdev.konsist.architecture.assertarchitecture.architecture5.project.application.sample
+
+import com.lemonappdev.konsist.architecture.assertarchitecture.architecture5.project.domain.DomainFirstClass
+
+class ApplicationThirdClass(val sampleParameter: DomainFirstClass)
