+++ lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture5/project/domain/sample/DomainSecondClass.kt
@@ -1,0 +1,5 @@
+package com.lemonappdev.konsist.architecture.assertarchitecture.architecture5.project.domain.sample
+
+import com.lemonappdev.konsist.architecture.assertarchitecture.architecture5.project.domain.DomainFirstClass
+
+class DomainSecondClass(val sampleParameter: DomainFirstClass)
