+++ lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture5/project/infrastructure/sample/InfrastructureSecondClass.kt
@@ -1,0 +1,5 @@
+package com.lemonappdev.konsist.architecture.assertarchitecture.architecture5.project.infrastructure.sample
+
+import com.lemonappdev.konsist.architecture.assertarchitecture.architecture5.project.infrastructure.InfrastructureFirstClass
+
+class InfrastructureSecondClass(val sampleParameter: InfrastructureFirstClass)
