+++ lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture5/project/presentation/sample/PresentationSecondClass.kt
@@ -1,0 +1,5 @@
+package com.lemonappdev.konsist.architecture.assertarchitecture.architecture5.project.presentation.sample
+
+import com.lemonappdev.konsist.architecture.assertarchitecture.architecture5.project.presentation.PresentationFirstClass
+
+class PresentationSecondClass(val sampleParameter: PresentationFirstClass)
