+++ lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture5/project/presentation/sample/PresentationThirdClass.kt
@@ -1,0 +1,5 @@
+package com.lemonappdev.konsist.architecture.assertarchitecture.architecture5.project.presentation.sample
+
+import com.lemonappdev.konsist.architecture.assertarchitecture.architecture5.project.application.ApplicationFirstClass
+
+class PresentationThirdClass(val sampleParameter: ApplicationFirstClass)
