+++ lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/assertarchitecture/architecture6/README.md
@@ -1,0 +1,3 @@
+# Architecture 6
+
+Test scenario where `EmptyLayer` has no files inside.
