--- lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/dependencyrules/circulardependency1/README.md
@@ -1,10 +1,0 @@
-# Circular dependency 1
-
-Test scenario with circular dependency:
-
-```mermaid
-%%{init: {'theme': 'forest'}}%%
-flowchart LR
-    Layer1 --> Layer2
-    Layer2 --> Layer1
-```
